Added functionality to compute centroids of bubbles
plminor.c:
Modified careaperim() to take a third flag and necessary arguments to compute the centroid of a single cell.
Created an arccentroid() function to compute the centroid of a circular
segment corresponding to a given arc.
Created a wrapper function centroids() to use careaperim() to compute the centroid of each cell.

plio.c:
Created a printcentroids() function to call centroids() and handle the
output files for the cetroids.
Created an information flag to optionally output the centroids to a file.

plequil.c,plminor.c:
Modified existing calls to careaperim() to take the updated number of
arguments.

xplat.c,plat.c,include.h:
Added necessary arrays and info titles.